+++ Blurred-ios/Recording Related Views/UploadDetailsViewController.swift
@@ -49,7 +49,8 @@ class UploadDetailsViewController: UIViewController {
         AVAsset(url: videoDetails).generateThumbnail { [weak self] (image) in
             DispatchQueue.main.async {
                 guard let image = image else { return }
-                self?.thumbnailView.image = image
+                let actualimage = image.rotate(radians: .pi / 2) 
+                self?.thumbnailView.image = actualimage
             }
         }
     }
@@ -128,5 +129,27 @@ extension AVAsset {
                 }
             })
         }
+    }
+}
+extension UIImage {
+    func rotate(radians: CGFloat) -> UIImage {
+        let rotatedSize = CGRect(origin: .zero, size: size)
+            .applying(CGAffineTransform(rotationAngle: CGFloat(radians)))
+            .integral.size
+        UIGraphicsBeginImageContext(rotatedSize)
+        if let context = UIGraphicsGetCurrentContext() {
+            let origin = CGPoint(x: rotatedSize.width / 2.0,
+                                 y: rotatedSize.height / 2.0)
+            context.translateBy(x: origin.x, y: origin.y)
+            context.rotate(by: radians)
+            draw(in: CGRect(x: -origin.y, y: -origin.x,
+                            width: size.width, height: size.height))
+            let rotatedImage = UIGraphicsGetImageFromCurrentImageContext()
+            UIGraphicsEndImageContext()
+
+            return rotatedImage ?? self
+        }
+
+        return self
     }
 }